Make chunk_group_content a regular function
